fix texture loading not working.
It turns out the texture origin is bottom left, not top left. There are
still too many hacks, but this is a start.

+++ src/Texture.h
@@ -20,7 +20,7 @@ public:
 	Texture(std::string);
 	virtual ~Texture();
 
-	Vec3f getColor(const int, const int) const;
+	Vec3f getColor(const float, const float) const;
 };
 
 #endif /* SRC_TEXTURE_H_ */
